R moved redis configuration to specific folder R added file for redisConnectionPool utils extension
