+++ Sources/Redis/Configuration/RedisConfiguration.swift
@@ -1,45 +1,19 @@
 import Foundation
-import NIOSSL
-import NIOPosix
 import Logging
 import NIOCore
+import NIOSSL
 import RediStack
 
 /// Configuration for connecting to a Redis instance
 public struct RedisConfiguration {
     public typealias ValidationError = RedisConnection.Configuration.ValidationError
 
-    public var serverAddresses: [SocketAddress]
-    public var password: String?
-    public var database: Int?
-    public var pool: PoolOptions
-    public var tlsConfiguration: TLSConfiguration?
-    public var tlsHostname: String?
-
-    public struct PoolOptions {
-        public var maximumConnectionCount: RedisConnectionPoolSize
-        public var minimumConnectionCount: Int
-        public var connectionBackoffFactor: Float32
-        public var initialConnectionBackoffDelay: TimeAmount
-        public var connectionRetryTimeout: TimeAmount?
-        public var onUnexpectedConnectionClose: ((RedisConnection) -> Void)?
-
-        public init(
-            maximumConnectionCount: RedisConnectionPoolSize = .maximumActiveConnections(2),
-            minimumConnectionCount: Int = 0,
-            connectionBackoffFactor: Float32 = 2,
-            initialConnectionBackoffDelay: TimeAmount = .milliseconds(100),
-            connectionRetryTimeout: TimeAmount? = nil,
-            onUnexpectedConnectionClose: ((RedisConnection) -> Void)? = nil
-        ) {
-            self.maximumConnectionCount = maximumConnectionCount
-            self.minimumConnectionCount = minimumConnectionCount
-            self.connectionBackoffFactor = connectionBackoffFactor
-            self.initialConnectionBackoffDelay = initialConnectionBackoffDelay
-            self.connectionRetryTimeout = connectionRetryTimeout
-            self.onUnexpectedConnectionClose = onUnexpectedConnectionClose
-        }
-    }
+    public let serverAddresses: [SocketAddress]
+    public let password: String?
+    public let database: Int?
+    public let pool: PoolOptions
+    public let tlsConfiguration: TLSConfiguration?
+    public let tlsHostname: String?
 
     public init(url string: String, tlsConfiguration: TLSConfiguration? = nil, pool: PoolOptions = .init()) throws {
         guard let url = URL(string: string) else { throw ValidationError.invalidURLString }
@@ -51,6 +25,7 @@ public struct RedisConfiguration {
             let scheme = url.scheme,
             !scheme.isEmpty
         else { throw ValidationError.missingURLScheme }
+
         guard scheme == "redis" || scheme == "rediss" else { throw ValidationError.invalidURLScheme }
         guard let host = url.host, !host.isEmpty else { throw ValidationError.missingURLHost }
 
@@ -80,7 +55,9 @@ public struct RedisConfiguration {
         database: Int? = nil,
         pool: PoolOptions = .init()
     ) throws {
-        if database != nil && database! < 0 { throw ValidationError.outOfBoundsDatabaseID }
+        if let database, database < 0 {
+            throw ValidationError.outOfBoundsDatabaseID
+        }
 
         try self.init(
             serverAddresses: [.makeAddressResolvingHost(hostname, port: port)],
@@ -106,26 +83,5 @@ public struct RedisConfiguration {
         self.tlsHostname = tlsHostname
         self.database = database
         self.pool = pool
-    }
-}
-
-extension RedisConnectionPool.Configuration {
-    internal init(_ config: RedisConfiguration, defaultLogger: Logger, customClient: ClientBootstrap?) {
-        self.init(
-            initialServerConnectionAddresses: config.serverAddresses,
-            maximumConnectionCount: config.pool.maximumConnectionCount,
-            connectionFactoryConfiguration: .init(
-                connectionInitialDatabase: config.database,
-                connectionPassword: config.password,
-                connectionDefaultLogger: defaultLogger,
-                tcpClient: customClient
-            ),
-            minimumConnectionCount: config.pool.minimumConnectionCount,
-            connectionBackoffFactor: config.pool.connectionBackoffFactor,
-            initialConnectionBackoffDelay: config.pool.initialConnectionBackoffDelay,
-            connectionRetryTimeout: config.pool.connectionRetryTimeout,
-            onUnexpectedConnectionClose: config.pool.onUnexpectedConnectionClose,
-            poolDefaultLogger: defaultLogger
-        )
     }
 }
